ai-service/requirements.txt
# # AI Service Requirements
# # LLM Test Script Dependencies

# # Groq API Client
# groq==0.11.0

# # Environment variables management
# python-dotenv==1.0.0

# # Basic Python libraries (usually included but explicitly listed for clarity)
# # os, glob, base64, datetime are built-in modules

# # Optional: Additional useful libraries for AI service
# numpy==1.24.3
# pillow==10.0.1

# # Development and testing
# pytest==7.4.3
# pytest-asyncio==0.21.1

# # FastAPI related (for future API development)
# fastapi==0.104.1
# uvicorn[standard]==0.24.0
# pydantic==2.5.0
# python-multipart==0.0.6

# # AI/ML libraries (for future model development)
# torch==2.1.0
# transformers==4.35.0
# scikit-learn==1.3.0
# pandas==2.0.3

# # Database connectivity (for future integration)
# pymongo==4.6.0
# redis==5.0.1

# # AWS integration (for future S3 usage)
# boto3==1.34.0

# # Security
# python-jose[cryptography]==3.3.0
# passlib[bcrypt]==1.7.4


groq
python-dotenv
tqdm